moved service runner definition to avoid naming conflicts
