+++ cobald/daemon/__init__.py
@@ -1,4 +1,4 @@
-from .service import ServiceRunner, service
+from .runners.service import ServiceRunner, service
 
 #: The runner invoked on daemon startup
 runtime = ServiceRunner()
